Fix user-echo-server hanging issue
On return of epoll_wait(), we go checking if the I/O-ready fd is
the fd that we are listening, if indeed, we use accept() to
get client fd. The potential issue here is, if we don't keep
checking whether the listener fd is still available (has pending
connections), we may get hanged next time we invoke epoll_wait(),
this is because that we didn't process all pending connection
requests before we invoke epoll_wait() again.

Such issue may occur if connection requests coming to the
server are too much in a short time, since the current impl only
checks the listener fd for once before invoking epoll_wait() again.

There is a solution to solve such issue, which is keep checking the
listener fd until it (accept()) returns EWOULDBLOCK. The trade-off
of this solution is, connected clients may wait for longer to get
their response, because we always service new connections coming in
a short time first.