Adjust TCP backlog of user-echo-server
In current impl, TCP backlog of the listener socket is 1, this
will cause some of requests get timeout error since backlog of
1 throttles the throughput of the server.

As listen(2) stated:
=====
If the backlog argument is greater than the value in
/proc/sys/net/core/somaxconn, then it is silently truncated to
that value.  Since Linux 5.4, the default in this file is 4096;
in earlier kernels, the default value is 128.
=====

Consider Linux 5.4 hasn't spreads to stable version of various
distros, I think using 128 as the backlog is appropriate.