Reject module insertion after kthread_run fail
In the original module init function, if the kthread_run fails, the
socket will be closed, but the module will be successfully inserted
into kernel. This behavior is unexpected and will cause kernel error
when the module exits. Therefore, this change let it return the error
code when the error happens.